requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pandas==2.1.3
tenacity==8.2.3

# LLM Providers (install what you need)
google-generativeai==0.3.2  # Free tier available
anthropic==0.8.0            # Free tier available
openai==1.6.0               # Paid

# Search APIs
google-search-results==2.4.2  # SerpAPI - has free tier

# Optional: Email enrichment
# pyhunter==2.1.0  # Uncomment if using Hunter.io

# Caching
diskcache==5.6.3

# Logging and utilities
colorlog==6.8.0
tqdm==4.66.1

# Data validation
validators==0.22.0

# Rate limiting
ratelimit==2.2.1
